requirements_dashboard.txt
# MLGenie Dashboard Requirements
# Additional packages needed for the enhanced dashboard functionality

# Database connectivity
mysql-connector-python>=8.0.0

# Enhanced UI components
streamlit-option-menu>=0.3.0
plotly>=5.0.0

# Data visualization and processing
pandas>=1.3.0
numpy>=1.21.0

# Machine learning (if not already installed)
scikit-learn>=1.0.0

# For progress tracking and async operations
streamlit-autorefresh>=0.0.6

# For better JSON handling
jsonschema>=4.0.0

# For file handling and utilities
pathlib2>=2.3.0

# For enhanced metrics and monitoring
psutil>=5.8.0  # For system resource monitoring
